move facet computation up and simplify
