+++ lib/utils.js
@@ -5,6 +5,7 @@ const semver = require('semver')
 const EntityCoordinates = require('./entityCoordinates')
 const ResultCoordinates = require('./resultCoordinates')
 const moment = require('moment')
+const { set } = require('lodash')
 
 function toResultCoordinatesFromRequest(request) {
   return new ResultCoordinates(
@@ -44,9 +45,31 @@ function extractDate(dateAndTime) {
   return moment(dateAndTime).format('YYYY-MM-DD')
 }
 
+function setIfValue(target, path, value) {
+  if (!value) return
+  set(target, path, value)
+}
+
+function setToArray(values) {
+  const result = Array.from(values)
+    .filter(e => e)
+    .sort()
+  return result.length === 0 ? null : result
+}
+
+function addArrayToSet(array, set, valueExtractor) {
+  if (!array || !array.length) return set
+  valueExtractor = valueExtractor || (value => value)
+  for (let entry of array) set.add(valueExtractor(entry))
+  return set
+}
+
 module.exports = {
   toEntityCoordinatesFromRequest,
   toResultCoordinatesFromRequest,
   getLatestVersion,
-  extractDate
+  extractDate,
+  setIfValue,
+  setToArray,
+  addArrayToSet
 }
